Factor `ct` module into `primitives`
It only contained `u32_max` and `u32_min` at this point, so this moves
them to `primitives` with similar low-level free functions
