--- src/ct.rs
@@ -1,33 +1,0 @@
-pub use ctutils::Choice as ConstChoice;
-pub use ctutils::CtOption as ConstCtOption;
-
-/// `const` equivalent of `u32::max(a, b)`.
-pub const fn u32_max(a: u32, b: u32) -> u32 {
-    ConstChoice::from_u32_lt(a, b).select_u32(a, b)
-}
-
-/// `const` equivalent of `u32::min(a, b)`.
-pub const fn u32_min(a: u32, b: u32) -> u32 {
-    ConstChoice::from_u32_lt(a, b).select_u32(b, a)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{u32_max, u32_min};
-
-    #[test]
-    fn test_u32_const_min() {
-        assert_eq!(u32_min(0, 5), 0);
-        assert_eq!(u32_min(7, 0), 0);
-        assert_eq!(u32_min(7, 5), 5);
-        assert_eq!(u32_min(7, 7), 7);
-    }
-
-    #[test]
-    fn test_u32_const_max() {
-        assert_eq!(u32_max(0, 5), 5);
-        assert_eq!(u32_max(7, 0), 7);
-        assert_eq!(u32_max(7, 5), 7);
-        assert_eq!(u32_max(7, 7), 7);
-    }
-}
