+++ src/modular/bingcd/xgcd.rs
@@ -1,7 +1,7 @@
 //! The Binary Extended GCD algorithm.
 use super::gcd::bingcd_step;
-use crate::ct::u32_max;
 use crate::modular::bingcd::matrix::{DividedIntMatrix, DividedPatternMatrix, PatternMatrix, Unit};
+use crate::primitives::u32_max;
 use crate::{ConstChoice, Int, Limb, NonZeroUint, Odd, OddUint, U64, U128, Uint, Word};
 
 /// Binary XGCD update step.
