+++ src/uint/gcd.rs
@@ -1,8 +1,8 @@
 //! This module implements Binary (Extended) GCD for [`Uint`].
 
-use crate::ct::u32_min;
 use crate::modular::bingcd::xgcd::PatternXgcdOutput;
 use crate::modular::safegcd;
+use crate::primitives::u32_min;
 use crate::{ConstChoice, Gcd, Int, NonZero, NonZeroUint, Odd, OddUint, Uint, Xgcd};
 
 impl<const LIMBS: usize> Uint<LIMBS> {
